requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile
#
alembic==1.12.1
    # via -r requirements.in
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   starlette
    #   watchfiles
click==8.1.7
    # via uvicorn
exceptiongroup==1.1.3
    # via anyio
faker==19.13.0
    # via -r requirements.in
fastapi[asunio]==0.104.1
    # via -r requirements.in
greenlet==3.0.1
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
httptools==0.6.1
    # via uvicorn
idna==3.4
    # via anyio
importlib-metadata==6.8.0
    # via alembic
importlib-resources==6.1.0
    # via alembic
mako==1.2.4
    # via alembic
markupsafe==2.1.3
    # via mako
psycopg2-binary==2.9.9
    # via -r requirements.in
pydantic==2.4.2
    # via fastapi
pydantic-core==2.10.1
    # via pydantic
python-dateutil==2.8.2
    # via faker
python-dotenv==1.0.0
    # via uvicorn
pyyaml==6.0.1
    # via uvicorn
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via anyio
sqlalchemy==2.0.23
    # via
    #   -r requirements.in
    #   alembic
starlette==0.27.0
    # via fastapi
typing-extensions==4.8.0
    # via
    #   alembic
    #   annotated-types
    #   faker
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   uvicorn
uvicorn[standard]==0.23.2
    # via -r requirements.in
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
zipp==3.17.0
    # via
    #   importlib-metadata
    #   importlib-resources
